update: rename Header to Hero
