+++ infrastructure/eid-wallet/src/routes/(auth)/e-passport/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import { Header } from "$lib/fragments";
+import { Hero } from "$lib/fragments";
 import IdentityCard from "$lib/fragments/IdentityCard/IdentityCard.svelte";
 import { ButtonAction } from "$lib/ui";
 
@@ -9,7 +9,7 @@ const handleFinish = async () => {};
 
 <main class="h-[max-content] pt-[5.2vh] px-[5vw] pb-[4.5vh] flex flex-col justify-between">
     <section>
-        <Header
+        <Hero
             title="Your ePassport and eVault are ready"
             subtitle="Log into any W3DS platform without passwords. It’s tied to this phone; if lost, you’ll need to revoke and reissue it on a new device."
             class="mb-2"
